+++ src/models/data.js
@@ -1,0 +1,5 @@
+// List contain all application Data.
+var Lists = [];
+
+
+exports.Lists;
